// library/ng-admin-lte/src/lib/model/alt-color.ts
export enum AltColor {
  DEFAULT = 'default',
  PRIMARY = 'primary',
  INFO = 'aqua',
  SUCCESS = 'green',
  WARNING = 'yellow',
  DANGER = 'red',
}

/**
 * support background colors
 */
export enum AltBgColor {
  red = 'red',
  yellow = 'yellow',
  aqua = 'aqua',
  blue = 'blue',
  light_blue = 'light-blue',
  green = 'green',
  navy = 'navy',
  teal = 'teal',
  olive = 'olive',
  lime = 'lime',
  orange = 'orange',
  fuchsia = 'fuchsia',
  purple = 'purple',
  maroon = 'maroon',
  black = 'black',
  red_active = 'red-active',
  yellow_active = 'yellow-active',
  aqua_active = 'aqua-active',
  blue_active = 'blue-active',
  light_blue_active = 'light-blue-active',
  green_active = 'green-active',
  navy_active = 'navy-active',
  teal_active = 'teal-active',
  olive_active = 'olive-active',
  lime_active = 'lime-active',
  orange_active = 'orange-active',
  fuchsia_active = 'fuchsia-active',
  purple_active = 'purple-active',
  maroon_active = 'maroon-active',
  black_active = 'black-active',
}

export enum AltButtonColor {
  DEFAULT = 'default',
  PRIMARY = 'primary',
  INFO = 'info',
  SUCCESS = 'success',
  WARNING = 'warning',
  DANGER = 'danger',
}

export enum AltColorName {
  PRIMARY = 'Primary',
  INFO = 'Info',
  SUCCESS = 'Success',
  WARNING = 'Warning',
  DANGER = 'Danger',
  GRAY = 'Gray',
  NAVY = 'Navy',
  TEAL = 'Teal',
  PURPLE = 'Purple',
  ORANGE = 'Orange',
  MAROON = 'Maroon',
  BLACK = 'Black',
}